% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{cor.emp.int}
\alias{cor.emp.int}
\title{cor.emp.int}
\usage{
cor.emp.int(zeta, nChainFit, Xt, parMargin, typeMargin)
}
\arguments{
\item{zeta}{correlation of Gaussian multivariates}

\item{nChainFit}{number of simulated variates}

\item{Xt}{simulated occurrences, n x 2 matrix}

\item{parMargin}{parameters of the margins 2 x 3}

\item{typeMargin}{type of marginal distribution: 'EGPD' or 'mixExp'}
}
\value{
\item{scalar}{correlation between simulated intensities}
}
\description{
Finds observed correlations between intensities corresponding
to a degree of correlation of Gaussian multivariate random numbers
}
\author{
Guillaume Evin
}
