% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{cor.emp.occ}
\alias{cor.emp.occ}
\title{cor.emp.occ}
\usage{
cor.emp.occ(w, Qtrans.mat, mat.comb, nLag, nChainFit, myseed = 1)
}
\arguments{
\item{w}{correlation of Gaussian multivariates}

\item{Qtrans.mat}{transition probabilities, 2 x ncomb matrix}

\item{mat.comb}{matrix of logical: ncomb x nlag}

\item{nLag}{order of the Markov chain}

\item{nChainFit}{number of simulated variates}

\item{myseed}{seed of random variates}
}
\value{
\item{scalar}{correlation between occurrences}
}
\description{
Finds observed correlations between occurrences corresponding
to a degree of correlation of Gaussian multivariate random numbers
}
\author{
Guillaume Evin
}
