% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{cor.obs.occ}
\alias{cor.obs.occ}
\title{cor.obs.occ}
\usage{
cor.obs.occ(pi00, pi0, pi1)
}
\arguments{
\item{pi00}{joint probability of having dry states}

\item{pi0}{probability of having a dry state}

\item{pi1}{probability of having a wet state}
}
\value{
\item{scalar}{matrix of observed correlations}
}
\description{
provide observed correlations between occurrences for all pairs of stations
see Mhanna et al. (2012)
}
\references{
Mhanna, Muamaraldin, and Willy Bauwens. “A Stochastic Space-Time Model 
for the Generation of Daily Rainfall in the Gaza Strip.” International Journal of 
Climatology 32, no. 7 (June 15, 2012): 1098–1112. doi:10.1002/joc.2305.
}
\author{
Guillaume Evin
}
