% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyPrecipGWEX-data.r
\docType{data}
\name{dailyPrecipGWEX}
\alias{dailyPrecipGWEX}
\title{daily observations of precipitation data}
\format{matrix of Observed precipitation: 3652 days x 3 stations}
\usage{
data(dailyPrecipGWEX)
}
\description{
Example of daily observations of precipitation (mm) for three fictive stations, for a period of ten years.
}
\references{
Evin, G., A.-C. Favre, and B. Hingray. 2018. “Stochastic Generation of Multi-Site Daily Precipitation Focusing on Extreme Events".
 Hydrol. Earth Syst. Sci. 22 (1): 655–672.
}
\author{
Guillaume Evin \email{guillaume.evin@irstea.fr}
}
\keyword{datasets}
