% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyTemperGWEX-data.r
\docType{data}
\name{dailyTemperGWEX}
\alias{dailyTemperGWEX}
\title{daily observations of temperature data}
\format{matrix of Observed temperature: 3652 days x 3 stations}
\usage{
data(dailyTemperGWEX)
}
\description{
Example of daily observations of temperature (mm) for three fictive stations, for a period of ten years.
}
\references{
Evin G., A.C. Favre, and B. Hingray. 2018. Stochastic Generators of Multi Site Daily Temperature:
Comparison of Performances in Various Applications. Theoretical and Applied Climatology.
}
\author{
Guillaume Evin \email{guillaume.evin@irstea.fr}
}
\keyword{datasets}
