% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{disag.3D.to.1D}
\alias{disag.3D.to.1D}
\title{disag.3D.to.1D}
\usage{
disag.3D.to.1D(Yobs, YObsAgg, mObsAgg, YSimAgg, mSimAgg, prob.class)
}
\arguments{
\item{Yobs}{matrix of observed intensities at 24h: (nTobs*3) x nStation}

\item{YObsAgg}{matrix of observed 3-day intensities: nTobs x nStation}

\item{mObsAgg}{vector of season corresponding to YobsAgg}

\item{YSimAgg}{matrix of simulated intensities per 3-day period: nTsim x nStation}

\item{mSimAgg}{vector of season corresponding to the period simulated}

\item{prob.class}{vector of probabilities indicating class of "similar" mean intensities}
}
\value{
\item{list}{Ysim matrix of disagregated daily precipitation, codeDisag matrix of disagregation codes}
}
\description{
disag.3D.to.1D
}
\author{
Guillaume Evin
}
