% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{dist.functions.EGPD.GI}
\alias{dist.functions.EGPD.GI}
\alias{dEGPD.GI}
\alias{pEGPD.GI}
\alias{qEGPD.GI}
\alias{rEGPD.GI}
\title{dEGPD.GI, pEGPD.GI, qEGPD.GI, rEGPD.GI}
\usage{
dEGPD.GI(x, kappa, sig, xi)

pEGPD.GI(x, kappa, sig, xi)

qEGPD.GI(p, kappa, sig, xi)

rEGPD.GI(n, kappa, sig, xi)
}
\arguments{
\item{x}{Vector of quantiles}

\item{kappa}{transformation parameter greater than 0}

\item{sig}{Scale parameter}

\item{xi}{Shape parameter}

\item{p}{Vector of probabilities}

\item{n}{Number of observations}
}
\value{
dEGPD.GI gives the density function, pEGPD.GI gives the distribution function, 
qEGPD.GI gives the quantile function, and rEGPD.GI generates random deviates.
}
\description{
Density function, distribution function, quantile function, random generation
for the unified EGPD distribution
}
\author{
Guillaume Evin
}
