% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{EGPD.GI.fit.PWM}
\alias{EGPD.GI.fit.PWM}
\title{EGPD.GI.fit.PWM}
\usage{
EGPD.GI.fit.PWM(x, xi = 0.05)
}
\arguments{
\item{x}{vector of parameters kappa,sig}

\item{xi}{shape parameter}
}
\value{
estimated parameters kappa, sig, xi
}
\description{
Parameter estimation of the unified EGPD distribution with the PWM method.
Numerical solver of the system of nonlinear equations
}
\author{
Guillaume Evin
}
