% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{EGPD.GI.fPWM}
\alias{EGPD.GI.fPWM}
\title{EGPD.GI.fPWM}
\usage{
EGPD.GI.fPWM(par, pwm, xi)
}
\arguments{
\item{par}{vector of parameters kappa,sig,xi}

\item{pwm}{set of probability weighted moments of order 0, 1 and 2}

\item{xi}{shape parameter}
}
\value{
differences between expected and target weighted moments
}
\description{
Parameter estimation of the unified EGPD distribution with the PWM method.
Set of equations which have to be equal to zero
}
\author{
Guillaume Evin
}
