% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{find.omega}
\alias{find.omega}
\title{find.omega}
\usage{
find.omega(rho.emp, Qtrans.mat, mat.comb, nLag, nChainFit)
}
\arguments{
\item{rho.emp}{target correlation between occurences}

\item{Qtrans.mat}{transition probabilities, 2 x ncomb matrix}

\item{mat.comb}{matrix of logical: ncomb x nlag}

\item{nLag}{order of the Markov chain}

\item{nChainFit}{length of the simulated chains used during the fitting}
}
\value{
\item{scalar}{needed correlation}
}
\description{
finds the correlation between normal variates leading to correlation between occurrences
}
\author{
Guillaume Evin
}
