% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{find.zeta}
\alias{find.zeta}
\title{find.zeta}
\usage{
find.zeta(eta.emp, nChainFit, Xt, parMargin, typeMargin)
}
\arguments{
\item{eta.emp}{target correlation between intensities}

\item{nChainFit}{number of simulations}

\item{Xt}{simulated occurrences, n x 2 matrix}

\item{parMargin}{parameters of the margins 2 x 3}

\item{typeMargin}{type of marginal distribution: 'EGPD' or 'mixExp'}
}
\value{
\item{scalar}{needed correlation}
}
\description{
finds the correlation between normal variates leading to correlation between intensities
}
\author{
Guillaume Evin
}
