% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{fit.copula.amount}
\alias{fit.copula.amount}
\title{fit.copula.amount}
\usage{
fit.copula.amount(P.mat, isPeriod, th, copulaInt, M0)
}
\arguments{
\item{P.mat}{precipitation matrix}

\item{isPeriod}{vector of logical n x 1 indicating the days concerned by a 3-month period}

\item{th}{threshold above which we consider that a day is wet (e.g. 0.2 mm)}

\item{copulaInt}{type of dependence between inter-site amounts: 'Gaussian' or 'Student'}

\item{M0}{covariance matrix of gaussianized prec. amounts for all pairs of stations}
}
\value{
\item{list}{list of estimates (e.g., M0, dfStudent)}
}
\description{
estimate parameters which control the spatial dependence between intensities using a copula
}
\author{
Guillaume Evin
}
