% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{fit.margin.cdf}
\alias{fit.margin.cdf}
\title{fit.margin.cdf}
\usage{
fit.margin.cdf(P.mat, isPeriod, th, type = c("EGPD", "mixExp"))
}
\arguments{
\item{P.mat}{precipitation matrix}

\item{isPeriod}{vector of logical n x 1 indicating the days concerned by a 3-month period}

\item{th}{threshold above which we consider that a day is wet (e.g. 0.2 mm)}

\item{type}{distribution: 'EGPD' or 'mixExp'}
}
\value{
\item{matrix}{matrix of estimates p x 3}
}
\description{
estimate parameters which control the marginal distribution of precipitation amounts
}
\author{
Guillaume Evin
}
