% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{functions.EGPD.GI}
\alias{functions.EGPD.GI}
\alias{EGPD.pGI}
\alias{EGPD.dGI}
\alias{EGPD.qGI}
\title{EGPD.pGI, EGPD.dGI, EGPD.qGI}
\usage{
EGPD.pGI(v, kappa)

EGPD.dGI(v, kappa)

EGPD.qGI(p, kappa)
}
\arguments{
\item{v}{probability}

\item{kappa}{transformation parameter greater than 0}

\item{p}{probability}
}
\value{
distribution, density and quantile of EGPD
}
\description{
First parametric family for G(v) = v^kappa: distribution, density and quantile function
}
\author{
Guillaume Evin
}
