% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{get.df.Student}
\alias{get.df.Student}
\title{get.df.Student}
\usage{
get.df.Student(P, Sig, max.df = 20)
}
\arguments{
\item{P}{matrix of non-zero precipitation (zero precipitation are set to NA)}

\item{Sig}{correlation matrix}

\item{max.df}{maximum degrees of freedom tested (default=20)}
}
\value{
nu estimate
}
\description{
Estimates the nu parameter (degrees of freedom) of the multivariate Student
distribution when the correlation matrix Sig is given
}
\references{
McNeil et al. (2005) "Quantitative Risk Management"
}
\author{
Guillaume Evin
}
