% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{get.emp.cdf.matrix}
\alias{get.emp.cdf.matrix}
\title{get.df.Student}
\usage{
get.emp.cdf.matrix(X)
}
\arguments{
\item{X}{matrix of positive precipitation}
}
\value{
matrix with cdf values (NA if zero precipitation)
}
\description{
get the cdf values (empirical distribution) of positive precipitation
}
\author{
Guillaume Evin
}
