% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{get.listOption}
\alias{get.listOption}
\title{get.listOption}
\usage{
get.listOption(listOption)
}
\arguments{
\item{listOption}{list containing fields corr. to the different options. Can be NULL if no options are set}
}
\value{
\item{listOption}{list of options}
}
\description{
get default options and check values proposed by the user
}
\author{
Guillaume Evin
}
