% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{get.M0}
\alias{get.M0}
\title{get.M0}
\usage{
get.M0(
  cor.obs,
  infer.mat.omega.out,
  nLag,
  parMargin,
  typeMargin,
  nChainFit,
  isParallel
)
}
\arguments{
\item{cor.obs}{matrix p x p of observed correlations between intensities for all pairs of stations}

\item{infer.mat.omega.out}{output of \code{\link{infer.mat.omega}}}

\item{nLag}{order of the Markov chain}

\item{parMargin}{parameters of the margins p x 3}

\item{typeMargin}{type of marginal distribution: 'EGPD' or 'mixExp'}

\item{nChainFit}{integer indicating the length of simulated chains}

\item{isParallel}{logical: indicate computation in parallel or not (easier for debugging)}
}
\value{
list with two items
\itemize{
  \item \strong{Xt} long simulation of the wet/dry states according to the model
  \item \strong{M0} covariance matrix of gaussianized prec. amounts for all pairs of stations
}
}
\description{
find matrix of correlations leading to estimates cor between intensities
}
\author{
Guillaume Evin
}
