% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{get.mat.omega}
\alias{get.mat.omega}
\title{get.mat.omega}
\usage{
get.mat.omega(cor.obs, Qtrans.mat, mat.comb, nLag, nChainFit, isParallel)
}
\arguments{
\item{cor.obs}{matrix p x p of observed correlations between occurrences for all pairs of stations}

\item{Qtrans.mat}{transition probabilities, 2 x ncomb matrix}

\item{mat.comb}{matrix of logical: ncomb x nlag}

\item{nLag}{order of the Markov chain}

\item{nChainFit}{length of the simulated chains used during the fitting}

\item{isParallel}{logical: indicate computation in parallel or not (easier for debugging)}
}
\value{
\item{matrix}{omega correlations for all pairs of stations}
}
\description{
find omega correlation leading to estimates cor between occurrences
}
\author{
Guillaume Evin
}
