% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{get.period.fitting.month}
\alias{get.period.fitting.month}
\title{get.period.fitting.month}
\usage{
get.period.fitting.month(m.char)
}
\arguments{
\item{m.char}{3-letter name of a month (e.g. 'JAN')

return the 3 indices corresponding to the 3-month period of a month ('JAN')}
}
\description{
get.period.fitting.month
}
