% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{get.vec.autocor}
\alias{get.vec.autocor}
\title{get.vec.autocor}
\usage{
get.vec.autocor(vec.ar1.obs, Xt, parMargin, typeMargin, nChainFit, isParallel)
}
\arguments{
\item{vec.ar1.obs}{vector of observed autocorrelations for all stations}

\item{Xt}{simulated occurrences given model parameters of wet/dry states}

\item{parMargin}{parameters of the margins 2 x 3}

\item{typeMargin}{type of marginal distribution: 'EGPD' or 'mixExp'}

\item{nChainFit}{integer indicating the length of the simulated chains}

\item{isParallel}{logical: indicate computation in parallel or not (easier for debugging)}
}
\value{
\item{vector}{vector of rho parameters to simulate the MAR process}
}
\description{
find rho autocorrelation leading to empirical estimates
}
\author{
Guillaume Evin
}
