% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWex_lib.r
\docType{class}
\name{Gwex-class}
\alias{Gwex-class}
\title{Class Gwex}
\description{
Defines a generic \code{\linkS4class{Gwex}} object.
GWex objects contain two slots:
- the version ('vX.X.X')
- the type of variable ('Prec' or 'Temp')
}
\author{
Guillaume Evin
}
