% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWex_lib.r
\docType{class}
\name{GwexFit-class}
\alias{GwexFit-class}
\title{Class GwexFit}
\description{
Defines a \code{\linkS4class{GwexFit}} object which is a \code{\linkS4class{Gwex}} object containing 'fit', a list containing the fitted parameters, and 'p', the number of stations.
See \link[GWEX]{fitGwexModel} for some examples.
}
\author{
Guillaume Evin
}
