% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWex_lib.r
\docType{class}
\name{GwexSim-class}
\alias{GwexSim-class}
\title{Defines a \code{\linkS4class{GwexSim}} object which is a \code{\linkS4class{Gwex}} object containing 'sim', an array containing the simulations, and 'dates', a vector of dates.
See \link[GWEX]{simGwexModel} for some examples.}
\description{
Defines a \code{\linkS4class{GwexSim}} object which is a \code{\linkS4class{Gwex}} object containing 'sim', an array containing the simulations, and 'dates', a vector of dates.
See \link[GWEX]{simGwexModel} for some examples.
}
\author{
Guillaume Evin
}
