% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{infer.autocor.amount}
\alias{infer.autocor.amount}
\title{infer.autocor.amount}
\usage{
infer.autocor.amount(
  P.mat,
  pr.state,
  isPeriod,
  nLag,
  th,
  parMargin,
  typeMargin,
  nChainFit,
  isMAR,
  isParallel
)
}
\arguments{
\item{P.mat}{precipitation matrix}

\item{pr.state}{probabilities of transitions for a Markov chain with lag \code{p}.}

\item{isPeriod}{vector of logical n x 1 indicating the days concerned by a 3-month period}

\item{nLag}{order of he Markov chain for the transitions between dry and wet states (=2 by default)}

\item{th}{threshold above which we consider that a day is wet (e.g. 0.2 mm)}

\item{parMargin}{parameters of the margins 2 x 3}

\item{typeMargin}{'EGPD' (Extended GPD) or 'mixExp' (Mixture of Exponentials). 'EGPD' by default}

\item{nChainFit}{integer, length of the runs used during the fitting procedure. =100000 by default}

\item{isMAR}{logical value, do we apply a Autoregressive Multivariate Autoregressive model (order 1) =TRUE by default}

\item{isParallel}{logical: indicate computation in parallel or not (easier for debugging)}
}
\value{
\item{list}{list of estimates (e.g., M0, dfStudent)}
}
\description{
special case of \code{\link{infer.dep.amount}} where there is only one station
}
\author{
Guillaume Evin
}
