% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{joint.proba.occ}
\alias{joint.proba.occ}
\title{joint.proba.occ}
\usage{
joint.proba.occ(P, th)
}
\arguments{
\item{P}{matrix of precipitation}

\item{th}{threshold above which we consider that a day is wet (e.g. 0.2 mm)}
}
\value{
\item{list}{list of joint probabilities}
}
\description{
joint probabilities of occurrences for all pairs of stations
}
\author{
Guillaume Evin
}
