% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{lag.trans.proba.mat.byClass}
\alias{lag.trans.proba.mat.byClass}
\title{lag.trans.proba.mat.byClass}
\usage{
\method{lag}{trans.proba.mat.byClass}(mat.prec, isPeriod, th, nlag, dayScale)
}
\arguments{
\item{mat.prec}{matrix of precipitation}

\item{isPeriod}{vector of logical n x 1 indicating the days concerned by a 3-month period}

\item{th}{threshold above which we consider that a day is wet (e.g. 0.2 mm)}

\item{nlag}{number of lag days}

\item{dayScale}{time resolution}
}
\value{
\item{list}{list with one item per station, where each item is a
matrix nLag^2 x (nLag+1) of transition probability between dry/wet state.
 The first nLag columns indicate the wet/dry states for the previous nLag days}
}
\description{
Estimate the transition probabilites between wet and dry states, for nlag previous days,
for all stations
}
\author{
Guillaume Evin
}
