% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{lagTransProbaVector}
\alias{lagTransProbaVector}
\title{lagTransProbaVector}
\usage{
lagTransProbaVector(vec.prec, isPeriod, th, nlag)
}
\arguments{
\item{vec.prec}{vector nx1 of precipitation for one station}

\item{isPeriod}{vector of logical n x 1 indicating the days concerned by a 3-month period}

\item{th}{threshold above which we consider that a day is wet (e.g. 0.2 mm)}

\item{nlag}{number of lag days}
}
\value{
\item{matrix}{matrix nLag^2 x (nLag+1) of transition probability between dry/wet state.
 The first nLag columns indicate the wet/dry states for the previous nLag days}
}
\description{
Estimate the transition probabilities between wet and dry states, for nlag previous days,
for one station
}
\author{
Guillaume Evin
}
