% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{mask.GWex.Yt}
\alias{mask.GWex.Yt}
\title{mask.GWex.Yt}
\usage{
mask.GWex.Yt(Xt, Yt)
}
\arguments{
\item{Xt}{simulated occurrences}

\item{Yt}{simulated intensities}
}
\value{
\item{matrix}{matrix n x p of simulated precipitations}
}
\description{
Mask intensities where there is no occurrence
}
\author{
Guillaume Evin
}
