% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{modify.cor.matrix}
\alias{modify.cor.matrix}
\title{modify.cor.matrix}
\usage{
modify.cor.matrix(cor.matrix)
}
\arguments{
\item{cor.matrix}{possibly non-positive definite correlation matrix}
}
\value{
positive definite correlation matrix
}
\description{
Modify a non-positive definite correlation matrix in order to have a positive 
definite matrix
}
\references{
Rousseeuw, P. J. and G. Molenberghs. 1993. Transformation of non positive semidefinite
correlation matrices. Communications in Statistics: Theory and Methods 22(4):965-984.

Rebonato, R., & Jackel, P. (2000). The most general methodology to create a valid 
correlation matrix for risk management and option pricing purposes. J. Risk, 2(2), 17-26.
}
\author{
Guillaume Evin
}
