% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{month2season}
\alias{month2season}
\title{month2season}
\usage{
month2season(vecMonth)
}
\arguments{
\item{vecMonth}{a vector of months given as integers 1:12}
}
\description{
transform vector of months to seasons
}
\author{
Guillaume Evin
}
