% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myParPrecGWEX-data.r
\docType{data}
\name{myParPrecGWEX}
\alias{myParPrecGWEX}
\title{fitted object of a precipitation GWEX model}
\format{object of class \code{\linkS4class{GwexFit}}}
\usage{
data(myParPrecGWEX)
}
\description{
Example of a object return by fit.GWex.prec when a GWEX model for precipitation data is fitted
}
\references{
Evin, G., A.-C. Favre, and B. Hingray. 2018. “Stochastic Generation of Multi-Site Daily Precipitation Focusing on Extreme Events".
 Hydrol. Earth Syst. Sci. 22 (1): 655–672.
}
\author{
Guillaume Evin \email{guillaume.evin@irstea.fr}
}
\keyword{datasets}
