% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myParTempGWEX-data.r
\docType{data}
\name{myParTempGWEX}
\alias{myParTempGWEX}
\title{fitted object of a temperature GWEX model}
\format{object of class \code{\linkS4class{GwexFit}}}
\usage{
data(myParTempGWEX)
}
\description{
Example of a object return by fit.GWex.temp when a GWEX model for temperature data is fitted
}
\references{
Evin G., A.C. Favre, and B. Hingray. 2018. Stochastic Generators of Multi Site Daily Temperature:
Comparison of Performances in Various Applications. Theoretical and Applied Climatology.
}
\author{
Guillaume Evin \email{guillaume.evin@irstea.fr}
}
\keyword{datasets}
