% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{PWM.EGPD.GI}
\alias{PWM.EGPD.GI}
\alias{EGPD.GI.mu0}
\alias{EGPD.GI.mu1}
\alias{EGPD.GI.mu2}
\title{EGPD.GI.mu0, EGPD.GI.mu1, EGPD.GI.mu2}
\usage{
EGPD.GI.mu0(kappa, sig, xi)

EGPD.GI.mu1(kappa, sig, xi)

EGPD.GI.mu2(kappa, sig, xi)
}
\arguments{
\item{kappa}{transformation parameter greater than 0}

\item{sig}{Scale parameter}

\item{xi}{Shape parameter}
}
\value{
Probability Weighted Moments
}
\description{
Probability Weighted Moments of order 0, 1 and 2 of the unified EGPD distribution
}
\author{
Guillaume Evin
}
