% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{QtransMat2Array}
\alias{QtransMat2Array}
\title{QtransMat2Array}
\usage{
QtransMat2Array(n, p, Qtrans.mat)
}
\arguments{
\item{n}{matrix of precipitation}

\item{p}{number of stations}

\item{Qtrans.mat}{transition probabilities, 2 x ncomb matrix}
}
\value{
\item{array}{array of transition probabilities with dimension n x p x n.comb}
}
\description{
reshape Qtrans.mat to an array
}
\author{
Guillaume Evin
}
