% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{sim.GWex.occ}
\alias{sim.GWex.occ}
\title{sim.GWex.occ}
\usage{
sim.GWex.occ(objGwexFit, vecMonth)
}
\arguments{
\item{objGwexFit}{object of class GwexFit}

\item{vecMonth}{vector n x 1 of integers indicating the months}
}
\value{
\item{matrix of logical}{occurrences simulated}
}
\description{
generate boolean variates which describe the dependence
between intersite occurrence correlations and wet/dry persistence
}
\author{
Guillaume Evin
}
