% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{sim.GWex.prec.1it}
\alias{sim.GWex.prec.1it}
\title{sim.GWex.prec.1it}
\usage{
sim.GWex.prec.1it(objGwexFit, vecDates, myseed, objGwexObs, prob.class)
}
\arguments{
\item{objGwexFit}{object of class GwexFit}

\item{vecDates}{vector of continuous dates}

\item{myseed}{seed of the random generation, to be fixed if the results need to be replicated}

\item{objGwexObs}{optional: necessary if we need observations to simulate (e.g. disaggregation of 3-day periods)}

\item{prob.class}{vector of probabilities indicating class of "similar" mean intensities}
}
\value{
\item{matrix}{Precipitation simulated for the dates contained in vec.Dates at the different stations}
}
\description{
Simulate one scenario of precipitation from the GWex model
}
\author{
Guillaume Evin
}
