% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{sim.GWex.Yt.Pr.get.param}
\alias{sim.GWex.Yt.Pr.get.param}
\title{sim.GWex.Yt.Pr.get.param}
\usage{
sim.GWex.Yt.Pr.get.param(objGwexFit, iM)
}
\arguments{
\item{objGwexFit}{object of class GwexFit}

\item{iM}{integer indicating the month}
}
\value{
\item{list}{list of parameters}
}
\description{
get relevant parameters
}
\author{
Guillaume Evin
}
