% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{sim.GWex.Yt.Pr}
\alias{sim.GWex.Yt.Pr}
\title{sim.GWex.Yt.Pr}
\usage{
sim.GWex.Yt.Pr(objGwexFit, vecMonth)
}
\arguments{
\item{objGwexFit}{object of class GwexFit}

\item{vecMonth}{vector n x 1 of integer indicating the months}
}
\value{
\item{matrix}{matrix n x p of uniform dependent variates}
}
\description{
generate uniform variates which describe the dependence between intersite amount
correlations
}
\author{
Guillaume Evin
}
