% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{sim.GWex.Yt}
\alias{sim.GWex.Yt}
\title{sim.GWex.Yt}
\usage{
sim.GWex.Yt(objGwexFit, vecMonth, Yt.Pr)
}
\arguments{
\item{objGwexFit}{object of class GwexFit}

\item{vecMonth}{vector of integer indicating the months}

\item{Yt.Pr}{uniform variates describing dependence between inter-site amounts}
}
\value{
\item{matrix}{matrix n x p of simulated non-zero precipitation intensities}
}
\description{
Inverse PIT: from the probability space to the precipitation
space
}
\author{
Guillaume Evin
}
