% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{sim.Zt.MAR}
\alias{sim.Zt.MAR}
\title{sim.Zt.MAR}
\usage{
sim.Zt.MAR(PAR, copulaInt, Zprev, p)
}
\arguments{
\item{PAR}{parameters for this class}

\item{copulaInt}{'Gaussian' or 'Student'}

\item{Zprev}{previous Gaussian variate}

\item{p}{number of stations}
}
\value{
\item{matrix}{matrix n x p of uniform dependent variates}
}
\description{
generate gaussian variates which describe the spatial and temporal dependence
between the sites (MAR(1) process)
}
\author{
Guillaume Evin
}
