% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{sim.Zt.Spatial}
\alias{sim.Zt.Spatial}
\title{sim.Zt.Spatial}
\usage{
sim.Zt.Spatial(PAR, copulaInt, p)
}
\arguments{
\item{PAR}{parameters for a class}

\item{copulaInt}{'Gaussian' or 'Student'}

\item{p}{number of stations}
}
\value{
\item{matrix}{matrix n x p of uniform dependent variates}
}
\description{
generate gaussian variates which describe the spatial dependence between the sites
}
\author{
Guillaume Evin
}
