% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{simPrecipOcc}
\alias{simPrecipOcc}
\title{simPrecipOcc}
\usage{
simPrecipOcc(nLag, n, pr)
}
\arguments{
\item{nLag}{order of the Markov chain}

\item{n}{integer indicating the length of simulated chains}

\item{pr}{vector of probabilies corr. to the conditional transition probabilities}
}
\value{
a vector Xt of length n with values 0/1 corr. to dry/wet states
}
\description{
find matrix of correlations leading to estimates cor between intensities
}
\author{
Guillaume Evin
}
