% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{unif.to.prec}
\alias{unif.to.prec}
\title{unif.to.prec}
\usage{
unif.to.prec(pI, typeMargin, U)
}
\arguments{
\item{pI}{vector of three parameters of the marginal distributions}

\item{typeMargin}{type of marginal distribution: 'EGPD' or 'mixExp'}

\item{U}{vector of uniform variates}
}
\value{
\item{matrix}{matrix of estimates p x 3}
}
\description{
from uniform variates to precipitation variates
}
\author{
Guillaume Evin
}
