% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWexPrec_lib.r
\name{wet.day.frequency}
\alias{wet.day.frequency}
\title{wet.day.frequency}
\usage{
wet.day.frequency(mat.prec, th)
}
\arguments{
\item{mat.prec}{matrix of precipitation (possibly for one month/period)}

\item{th}{threshold above which we consider that a day is wet (e.g. 0.2 mm)}
}
\value{
\item{vector of numeric}{wet day frequencies}
}
\description{
Estimate the wet day frequency (proportion of wet days) for all stations
}
\author{
Guillaume Evin
}
